requirements.txt
tkinter
requests
pytz
python-dotenv
logging
webbrowser
subprocess
csv
datetime
threading